+++ apps/starter-kit-next/next.config.mjs
@@ -1,0 +1,7 @@
+import { withGluestackUI } from "@gluestack-nightly/ui-next-adapter";
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  transpilePackages: []
+};
+
+export default withGluestackUI(nextConfig);