+++ packages/ui-next-adapter/CHANGELOG.md
@@ -1,5 +1,59 @@
 # @gluestack-nightly/ui-next-adapter
 
+## 1.0.36
+
+### Patch Changes
+
+- chore: removed gluestack-style and added nativewind and react-native-css-interop
+
+## 1.0.35
+
+### Patch Changes
+
+- fix: fixed next-adapter
+
+## 1.0.31
+
+### Patch Changes
+
+- chore: test
+
+## 1.0.26
+
+### Patch Changes
+
+- test
+
+## 1.0.7
+
+### Patch Changes
+
+- chore: test
+
+## 1.0.6
+
+### Patch Changes
+
+- test
+
+## 1.0.5
+
+### Patch Changes
+
+- fix: defined expo global constant **dev**
+
+## 1.0.4
+
+### Patch Changes
+
+- fix: defined global constants for expo **dev**
+
+## 1.0.3
+
+### Patch Changes
+
+- fix: added expo/html-elements and react-native-safe-area-context to transpile packages in ui-next-adapter
+
 ## 1.0.2
 
 ### Patch Changes